Added (some) support for multiple monitor setups and selective component loading at run-time.
This is a particularly messy commit. I moved a lot of code around and made a few small adjustments. To anyone reading this: I'm sorry.

Support for multiple monitors probably will not work as intended, because I do not have a second monitor set up to test with, but it should at least print warnings and I will be able to download the code on other computers to test.