Revert "Moved parameter definitions to a header file."
This reverts commit eacca317ca361089b5d23f6385e378809984bc30.

I really wish that worked but, as I keep forgetting, it does not. Static
is required in the header file and it breaks the application's logic for
allowing C.L.I. parameters to over-ride values assigned in the default
config file. Specifically, the value `register_param` assigns to "previous_load" does not persist; probably because the build script compiles as (static) libraries. Maybe I'll try working around this some day.